Added service classes for all entities and controllers + tests

+++ resource_service/src/main/java/com/example/resource_service/repository/ProductionOrderRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductionOrderRepo extends JpaRepository<ProductionOrder, Integer> {
+public interface ProductionOrderRepo extends JpaRepository<ProductionOrder, Long> {
 }
